EES-2328 Changes in response to PR comments
